Update CVE-2021_34527_PrintNightmare_Patch.ps1
We're changing the logic here a bit-- now that we want to patch everything unless specifically stated otherwise, we want to patch as long as $excludeFromMitigation is not 1. If the exclude mitigation EDF is checked in Automate, that value will always be 1 and otherwise will not sot his ensures it only skips mitigation when that box is checked.