OK.. So they DO something, just not exactly what you expect
`-EA 0` DOES keep `Get-Hotfix` from erroring out when `-Id` has a value but the hotfix isn't installed. If you provide it with `$Null` however, it still errors.

I still fixed it correctly, but I should not have removed the `-EA 0` because that is still necessary to get it to work in an `If` as a true/false value