Adjusted all occurrences of `$mitigationApplied`
Adjusted all occurrences of `$mitigationApplied=1` to be inside of `If (Test-MitigationApplied)` blocks because with the recent changes, whether mitigation is applied or not depends on whether `$excludeFromMitigation` is true and not always which path we end up in.

Also adjusted output message about mitigation being applied so that it's not lying when `$excludeFromMitigation` is true.